requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
qdrant-client==1.11.2
sentence-transformers==3.1.1
pymongo==4.9.1
requests==2.32.3
numpy==2.1.1
python-dotenv==1.0.1